Compile Rust code and call from R

rust_source() compiles and loads a single Rust file for use in R. rust_function() compiles and loads a single Rust function for use in R.

rust_source(
  file,
  code = NULL,
  dependencies = NULL,
  patch.crates_io = NULL,
  profile = c("dev", "release"),
  env = parent.frame(),
  cache_build = TRUE,
  quiet = FALSE
)

rust_function(code, env = parent.frame(), ...)

Arguments

file
Input rust file to source.

code
Input rust code, to be used instead of file.

dependencies
Character vector of dependencies lines to be added to the Cargo.toml file.

patch.crates_io
Character vector of patch statements for crates.io to be added to the Cargo.toml file.

profile
Rust profile. Can be either "dev" or "release". The default, "dev", compiles faster but produces slower code.

env
The R environment in which the wrapping functions will be defined.

cache_build
Logical indicating whether builds should be cached between calls to rust_source().

quiet
Logical indicating whether compile output should be generated or not.

...
Other parameters handed off to rust_source().

Value

The result from dyn.load(), which is an object of class DLLInfo. See getLoadedDLLs() for more details.

Examples

if (FALSE) {
# creating a single rust function
rust_function("fn add(a:f64, b:f64) -> f64 {a+b}")
add(2.5, 4.7)

# creating multiple rust functions at once
code <- r"(
use extendr_api::*;

#[extendr]
fn hello() -> &'static str {
    "Hello, world!"
}

#[extendr]
fn test( a: &str, b: i64) {
    println!("Data sent to Rust: {}, {}", a, b);
}
)"

rust_source(code = code)
hello()
test("a string", 42)
}